Composite Markup PNGs Now Assoicated by Book Title
Yay! Upon scribbling a markup, a UUID-4 is generated to serve as the base name for the respective `svg`and `jpg` file pairs that are visible in the `.kobo/markups` directory. When a markup is created, an `INSERT` query is also performed on the `KoboReader.sqlite` database. Within this insert, the aforementioned UUID-4 is inserted as the `Bookmark_ID` value. The following changes in this commit query the database, returning a corresponding `Volume_ID` that is then stripped down to the book title name. I use this book title name to save each newly created composite markup `png` to its respective book title.