experimenting with cropping page jpg
I'm seeing whether or not cropping the book pages to the svg bounding box (with trivial value of `50` as bounding box padding) is something worth exploring further. In light of the changes within this commit, I'm noticing 1.) cropping is great for underlines, but sucks for cases like margin notes,  and 2.) additional padding can result in black strips in the final output markup file in cases were the padding exceeds the extent of the book page.

I think cropping is an important step for being able to use python OCR libraries to create "normal" annotations from markup images. But its clear that there's a lot of ambiguity in users' intents when it comes markups.